# docs/swagger/swagger.yaml
basePath: /v1
definitions:
  contract.CreateDigitalAssistantRequest:
    properties:
      avatar:
        type: string
      code:
        type: string
      description:
        type: string
      name:
        type: string
      system_prompt:
        type: string
    required:
    - code
    - name
    type: object
...
  contract.DigitalAssistant:
    properties:
      avatar:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      org_id:
        type: integer
      owner_id:
        type: integer
      status:
        type: string
      system_prompt:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  contract.ListDigitalAssistantRequest:
    properties:
      keyword:
        type: string
      limit:
        type: integer
      list_all:
        type: boolean
      offset:
        type: integer
      status:
        type: string
    type: object
...
  dto.BaseResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  dto.CreateDigitalAssistantResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/contract.DigitalAssistant'
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      message:
        type: string
    type: object
...
  handler.DeleteDigitalAssistantRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  handler.GetDigitalAssistantRequest:
    properties:
      code:
        type: string
      id:
        type: integer
    type: object
...
  handler.UpdateDigitalAssistantRequest:
    properties:
      avatar:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      system_prompt:
        type: string
    required:
    - id
    type: object
  handler.UpdateDigitalAssistantStatusRequest:
    properties:
      id:
        type: integer
      status:
        type: string
    required:
    - id
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Leros 数字助手平台 API，提供数字助手管理、技能调用、事件处理等功能
  title: Leros API
  version: "1.0"
paths:
...
  /CreateDigitalAssistant:
    post:
      consumes:
      - application/json
      description: 创建一个新的数字助手实例
      parameters:
      - description: 创建数字助手请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contract.CreateDigitalAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.CreateDigitalAssistantResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 创建数字助手
      tags:
      - DigitalAssistant
...
  /DeleteDigitalAssistant:
    post:
      consumes:
      - application/json
      description: 根据ID删除数字助手
      parameters:
      - description: 删除数字助手请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.DeleteDigitalAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 删除数字助手
      tags:
      - DigitalAssistant
...
  /GetDigitalAssistant:
    post:
      consumes:
      - application/json
      description: 根据ID或Code获取数字助手详情
      parameters:
      - description: 获取数字助手请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.GetDigitalAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.CreateDigitalAssistantResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取数字助手详情
      tags:
      - DigitalAssistant
...
  /ListDigitalAssistant:
    post:
      consumes:
      - application/json
      description: 分页查询数字助手列表
      parameters:
      - description: 查询列表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contract.ListDigitalAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.CreateDigitalAssistantResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 查询数字助手列表
      tags:
      - DigitalAssistant
...
  /UpdateDigitalAssistant:
    post:
      consumes:
      - application/json
      description: 更新数字助手基本信息
      parameters:
      - description: 更新数字助手请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDigitalAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.CreateDigitalAssistantResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 更新数字助手
      tags:
      - DigitalAssistant
  /UpdateDigitalAssistantStatus:
    post:
      consumes:
      - application/json
      description: 更新数字助手的运行状态
      parameters:
      - description: 更新状态请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDigitalAssistantStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 更新数字助手状态
      tags:
      - DigitalAssistant
...
schemes:
- http
- https
swagger: "2.0"